Added files via upload
Pixelator: The tool is a faster alternative to the standard ray tracer. It takes an stl triangle mesh in the ascii format and returns the triangles that are visible from the given direction.

Voxelator: The tool is developed for quick estimation of aerodynamics and aerothermodynamics of re-entering objects. The tool uses state-of-the-art graphical technique of using voxel to create the deisred geometery. The geometry  is made using a combination of primitives (current version models only sphere, cube, cylinder, and cone). 

For more details about the tools, please read : Piyush M. Mehta, Gonzalo Blanco-Arnao, Davide Bonetti, et al., Computer Graphics for Space Debris, 6th International Conference on Astrodynamics Tools and Techniques, Darmstadt, Germany, 14-17 March, 2016.